divided the charts into separate js, divided section.css into separate css
