+++ src/styles/projects.module.css
@@ -1,0 +1,75 @@
+/*STYLED BUTTONS*/
+
+.buttons {
+    display: inline-flex;
+    align-items: center;
+    border: none;
+    background: transparent;
+    color: var(--darkbrown);
+    font-family: "Open Sans", sans-serif;
+    font-weight: bold;
+    font-size: 15px;
+    cursor: pointer;
+}
+
+.buttons:hover {
+    color: var(--yellow);
+    transition: 0.5s ease;
+}
+
+/*PROJECTS SECTION*/
+
+.project-grid {
+    width: 50%;
+    margin: auto;
+    display: grid;
+    grid-template-columns: 30% 70%;
+    column-gap: 15px;
+}
+
+.project-photos {
+    height: fit-content;
+    box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.2);
+    margin-top: 5px;
+    border-radius: 5px;
+}
+
+.project-item {
+    font-family: "Nanum Myeongjo", serif;
+    color: var(--darkbrown);
+    font-size: 19px;
+}
+
+.project-item h2 {
+    font-family: "Yeseva One", cursive;
+    color: var(--darkbrown);
+    font-size: 30px;
+    letter-spacing: 1px;
+    margin-bottom: 0;
+    margin-top: 0;
+}
+
+.project-item h3 {
+    font-family: "Yeseva One", cursive;
+    color: var(--lightbrown);
+    font-size: 23px;
+    margin: 0;
+}
+
+.project-Icons {
+    display: inline-flex;
+    flex-wrap: wrap;
+    gap: 12px;
+    flex-direction: row;
+    align-items: center;
+}
+
+.view-more {
+    height: fit-content;
+    margin-bottom: 38px;
+}
+
+.view-less {
+    height: 30vh;
+    margin-bottom: 38px;
+}
