--- src/styles/sections.module.css
@@ -1,305 +1,0 @@
-@import "./typography.css";
-
-/* CONTAINER */
-
-.section {
-    width: 100%;
-    min-height: 100vh;
-    position: relative;
-    overflow: hidden;
-}
-
-.divider {
-    height: 1px;
-    width: 40%;
-    background: rgba(0, 0, 0, 0.1);
-    margin: auto;
-    margin-top: 75px;
-    margin-bottom: 75px;
-}
-
-.buttons {
-    display: inline-flex;
-    align-items: center;
-    border: none;
-    background: transparent;
-    color: var(--darkbrown);
-    font-family: "Open Sans", sans-serif;
-    font-weight: bold;
-    font-size: 15px;
-    cursor: pointer;
-}
-
-.buttons:hover {
-    color: var(--yellow);
-    transition: 0.5s ease;
-}
-
-
-/*TYPOGRAPHY*/
-
-.highlight {
-    background: linear-gradient(
-        to bottom,
-        transparent 50%,
-        var(--yellow) 50%
-        );
-}
-
-a:hover {
-    transition: ease 0.5s;
-    opacity: 50%;
-}
-
-/*HEADINGS*/
-
-.heading {
-  font-family: 'Yeseva One', cursive;
-  text-align: center;
-  color: var(--darkbrown);
-  font-size: 52px;
-  letter-spacing: 3px;
-  margin: 0;
-  margin-bottom: 15px;
-}
-
-.heading-decoration {
-    font-family: 'Nanum Myeongjo', serif;
-    font-weight: bold;
-    text-align: center;
-    color: var(--darkbrown);
-    font-size: 25px;
-    margin: 0;
-    margin-top: 75px;
-}
-
-.subheading {
-    font-family: 'Yeseva One', cursive;
-    width: 50%;
-    margin: auto;
-    margin-bottom: 15px;
-    color: var(--darkbrown);
-    font-size: 25px;
-    letter-spacing: 5px;
-}
-
-/*ABOUT ME SECTION*/
-
-.profile {
-    width: 75%;
-    margin: auto;
-    margin-bottom: 50px;
-    text-align: center;
-    font-family: 'Nanum Myeongjo', serif;
-    font-size: 23px;
-    color: var(--darkbrown);
-}
-
-.profile-pic {
-    height: 225px;
-    width: 225px;
-    margin: auto;
-    background-image: url("../images/profile.webp");
-    background-size: cover;
-    background-repeat: no-repeat;
-    border-radius: 50%;
-    border: 7px solid var(--lightestyellow);
-    display: inline-block;
-}
-
-.education {
-    width: 50%;
-    margin: auto;
-    display: grid;
-    grid-template-columns: 10% 35% 55%;
-    column-gap: 15px;
-    margin-bottom: 5vh;
-}
-
-.education-item {
-    color: var(--darkbrown);
-    font-family: 'Nanum Myeongjo', serif;
-    font-size: 19px;
-}
-
-.education-item p1 {
-    font-weight: 800;
-    margin-top: 5px;
-    margin-bottom: 5px;
-}
-
-.education-item p {
-    color: var(--lightbrown);
-    margin: 0;
-}
-
-.skills {
-    position: relative;
-    display: flex;
-    width: 50%;
-    margin: auto;
-}
-
-.skills-item {
-    display: flex;
-    flex-direction: column;
-    width: auto;
-    color: var(--darkbrown);
-    font-family: 'Nanum Myeongjo', serif;
-    font-weight: 800;
-    font-size: 19px;
-    margin-right: 15px;
-}
-
-.skills-item p {
-    margin: 0;
-    margin-bottom: 15px;
-}
-
-.percentage {
-    display: flex;
-    flex-direction: column;
-    width: auto;
-    color: var(--lightbrown);
-    font-family: 'Nanum Myeongjo', serif;
-    font-size: 19px;
-    text-align: right;
-    margin-left: 15px;
-}
-
-.percentage p {
-    margin: 0;
-    margin-bottom: 15px;
-}
-
-.chart {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-}
-
-.chart div {
-    margin-bottom: 15px;
-}
-
-.bar-animation {
-    animation: 2s forwards slidein;
-    animation-play-state: paused;
-}
-
-@keyframes slidein {
-  from {
-    width: 0%;
-    height: 100%;
-    border-radius: 5px;
-  }
-
-  to {
-    height: 100%;
-    border-radius: 5px;
-  }
-}
-
-.language {
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-    align-items: left;
-    gap: 20px;
-    margin: auto;
-    margin-bottom: 15px;
-}
-
-.language p {
-    color: var(--darkbrown);
-    font-family: 'Nanum Myeongjo', serif;
-    font-weight: 800;
-    font-size: 19px;
-    margin: 0;
-}
-
-/*PROJECTS SECTION*/
-
-.project-grid {
-    width: 50%;
-    margin: auto;
-    display: grid;
-    grid-template-columns: 30% 70%;
-    column-gap: 15px;
-}
-
-.project-photos {
-    height: fit-content;
-    box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.2);
-    margin-top: 5px;
-    border-radius: 5px;
-}
-
-.project-item {
-    font-family: "Nanum Myeongjo", serif;
-    color: var(--darkbrown);
-    font-size: 19px;
-}
-
-.project-item h2 {
-    font-family: "Yeseva One", cursive;
-    color: var(--darkbrown);
-    font-size: 30px;
-    letter-spacing: 1px;
-    margin-bottom: 0;
-    margin-top: 0;
-}
-
-.project-item h3 {
-    font-family: "Yeseva One", cursive;
-    color: var(--lightbrown);
-    font-size: 23px;
-    margin: 0;
-}
-
-.project-Icons {
-    display: inline-flex;
-    flex-wrap: wrap;
-    gap: 12px;
-    flex-direction: row;
-    align-items: center;
-}
-
-.view-more {
-    height: fit-content;
-    margin-bottom: 38px;
-}
-
-.view-less {
-    height: 30vh;
-    margin-bottom: 38px;
-}
-
-/* CONTACT INFO*/
-
-.contact-grid {
-    width: 30%;
-    margin: auto;
-    display: grid;
-    grid-template-columns: 20% 80%;
-    grid-row-gap: 23px;
-    margin-bottom: 50px;
-}
-
-.contact-item {
-    margin: auto;
-    color: var(--darkbrown);
-    font-family: "Open Sans", sans-serif;
-    font-weight: bold;
-    font-size: 15px;
-}
-
-.contact-item:hover {
-    transform: translateY(-5px);
-    color: var(--lightbrown);
-    transition: 0.3s ease;
-}
-
-.disabled {
-    pointer-events: none;
-}
